requirements.txt
ansible~=4.7.0
boto>=2.49.0
botocore>=1.18.0
boto3>=1.15.0
awscli>=1.22.6
kubernetes>=12.0.0
backoff>=1.11.1
polling>=0.3.2
pyyaml>=6.0